requirements.txt
Flask
Werkzeug
PyPDF2
Pillow
EeasyPrint
redis
flask-redis
PyMuPdf
mammoth
weasyprint
python-docx
fpdf
reportlab